a variable called io to emit and receive events. i get it by requiring socket.io and passing the variable server to it,A socket is also passed to the callback which i can use to communicate with the client bi-directionally. for listening the events by writing socket.on('EVENT_NAME').for custom event i have used  event:new:image , it indicates connected user send some images
i have emit this event from browser when the user needs to share an image. The data parameter in the callback represents the data. It consists of two properties: sender and image. Once we have the data from one user we broadcast it to all other users except the one who sent it. All the connected users listen to this event in the client side so that they can update their view with the incoming image. i have emit the event by writing: socket.broadcast.emit('event:incoming:image',data).